pyproject.toml
[tool.poetry]
name = "tungsten-sds"
version = "0.6.2"
description = "An MSDS parser."
authors = ["Dennis Pham <dennis@dennispham.me>", "Brian Lu <me@greencappuccino.net>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/Den4200/tungsten"
keywords = ["sds", "parser"]
packages = [{ include = "tungsten" }]
include = ["tabula-1.0.6-SNAPSHOT-jar-with-dependencies.jar"]

[tool.poetry.dependencies]
"pdfminer.six" = "^20220524"
python = "^3.10"
tabula-py = "^2.5.1"
numpy = "^1.24.2"
pillow = "^9.4.0"
opencv-python-headless = "^4.7.0.68"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
taskipy = "^1.10.3"
pre-commit = "^2.20.0"

[tool.taskipy.tasks]
lint = "flake8 ."
pre-commit = "pre-commit install"

[tool.isort]
# See https://pycqa.github.io/isort/docs/configuration/multi_line_output_modes.html#3-vertical-hanging-indent
multi_line_output = 3

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
